Use standard newline at end of lines
This is instead of moving the terminal cursor position to where we
think it should be; the benefit of doing it this way is that there are
no problems (besides tabstops getting misaligned) if we encounter
characters that the terminal displays as wider than we expect. E.g.
imagine that someone tries to output a line of wide characters that
overflows to the next line; by moving the cursor to the start of that
line, some characters will be overwritten.